[Internal] DSPreview KeyInputs
The DSPreview counts as a spectator-user. Scripts in-game can read it and assign it a in-world player, however there is no code to read inputs, meaning the only way to move is exploiting a bug when closing the scoreboard or chat.

This adds proper code to pass inputs to the DSPreview client.